Changed some classes to interfaces and added new packages
