+++ src/bomberman/entity/staticobjects/Wall.java
